de-couple variable placeholder logics
