+++ bukkit/latest/src/main/java/io/github/wysohn/triggerreactor/bukkit/manager/trigger/share/api/placeholder/PlaceholderExpansionSupport.java
@@ -1,13 +1,12 @@
 package io.github.wysohn.triggerreactor.bukkit.manager.trigger.share.api.placeholder;
 
 import io.github.wysohn.triggerreactor.core.main.TriggerReactorCore;
-import io.github.wysohn.triggerreactor.core.manager.GlobalVariableManager;
-import io.github.wysohn.triggerreactor.core.script.interpreter.TemporaryGlobalVariableKey;
 import me.clip.placeholderapi.expansion.PlaceholderExpansion;
 import org.bukkit.entity.Player;
 
 public class PlaceholderExpansionSupport extends PlaceholderExpansion {
     private final TriggerReactorCore plugin;
+    private final IVariablePlaceholder variablePlaceholder;
 
     /**
      * Since we register the expansion inside our own plugin, we
@@ -18,6 +17,7 @@ public class PlaceholderExpansionSupport extends PlaceholderExpansion {
      */
     public PlaceholderExpansionSupport(TriggerReactorCore plugin) {
         this.plugin = plugin;
+        this.variablePlaceholder = new VariablePlaceholder(plugin);
     }
 
     /**
@@ -81,61 +81,8 @@ public class PlaceholderExpansionSupport extends PlaceholderExpansion {
         return plugin.getVersion();
     }
 
-    /**
-     * This is the method called when a placeholder with our identifier
-     * is found and needs a value.
-     * <br>We specify the value identifier in this method.
-     * <br>Since version 2.9.1 can you use OfflinePlayers in your requests.
-     *
-     * @param player     A Player who performed the task which contains interaction on
-     *                   PlaceholderAPI.
-     * @param identifier A String containing the identifier/value.
-     * @return possibly-null String of the requested identifier.
-     */
     @Override
-    public String onPlaceholderRequest(Player player, String identifier) {
-        if(identifier == null || identifier.length() == 0 || identifier.equals("?"))
-            return "";
-        if (identifier.toLowerCase().equals("version")) {
-            return plugin.getVersion();
-        }
-        //%tr_?<variable name>% - temporary global variable
-        if (identifier.startsWith("?")) {
-            String variableName = identifier.substring(1).replace('_', '.');
-            TemporaryGlobalVariableKey tempKey = new TemporaryGlobalVariableKey(variableName);
-            GlobalVariableManager.GlobalVariableAdapter adapter = (GlobalVariableManager.GlobalVariableAdapter) plugin.getVariableManager().getGlobalVariableAdapter();
-            Object value = adapter.get(tempKey);
-            if (value == null) {
-                return "";
-            }
-            if (value instanceof Number) {
-                value = String.valueOf(value);
-            }
-
-            if (!(value instanceof String)) {
-                return "";
-            } else {
-                String output = (String) value;
-                return output;
-            }
-        }
-
-        // %tr_<variable name>%
-        String variableName = identifier.replace('_', '.');
-        GlobalVariableManager vm = plugin.getVariableManager();
-        Object value = vm.get(variableName);
-        if (value == null) {
-            return "";
-        }
-        if (value instanceof Number) {
-            value = String.valueOf(value);
-        }
-
-        if (!(value instanceof String)) {
-            return "";
-        } else {
-            String output = (String) value;
-            return output;
-        }
+    public String onPlaceholderRequest(Player p, String params) {
+        return variablePlaceholder.onPlaceholderRequest(p, params);
     }
 }
